add return error when file not set

+++ DynaMind-GDALModules/scripts/GDALModules/dm_sftp_hotstart.py
@@ -190,6 +190,10 @@ class DM_Hoststart_SFTP(Module):
 
         def run(self):
             db = self.getGDALData("city")
+            if self.real_file_name == "":
+                log("File not downloaded", Error)
+                self.setStatus(MOD_EXECUTION_ERROR)
+                return
             db.setGDALDatabase(self.real_file_name)
             os.remove(self.real_file_name)
